fix : moving logic from repository layer to use case layeer

+++ src/Applications/use_case/DeleteAuthenticationUseCase.js
@@ -1,25 +1,32 @@
 class DeleteAuthenticationUseCase {
-  constructor({
-    authenticationRepository,
-  }) {
+  constructor({ authenticationRepository }) {
     this._authenticationRepository = authenticationRepository;
   }
 
   async execute(useCasePayload) {
     this._validatePayload(useCasePayload);
     const { refreshToken } = useCasePayload;
-    await this._authenticationRepository.checkAvailabilityToken(refreshToken);
+    const tokenCount =
+      await this._authenticationRepository.checkAvailabilityToken(refreshToken);
+    if (!tokenCount) {
+      const InvariantError = require("../../Commons/exceptions/InvariantError");
+      throw new InvariantError("refresh token tidak ditemukan di database");
+    }
     await this._authenticationRepository.deleteToken(refreshToken);
   }
 
   _validatePayload(payload) {
     const { refreshToken } = payload;
     if (!refreshToken) {
-      throw new Error('DELETE_AUTHENTICATION_USE_CASE.NOT_CONTAIN_REFRESH_TOKEN');
+      throw new Error(
+        "DELETE_AUTHENTICATION_USE_CASE.NOT_CONTAIN_REFRESH_TOKEN"
+      );
     }
 
-    if (typeof refreshToken !== 'string') {
-      throw new Error('DELETE_AUTHENTICATION_USE_CASE.PAYLOAD_NOT_MEET_DATA_TYPE_SPECIFICATION');
+    if (typeof refreshToken !== "string") {
+      throw new Error(
+        "DELETE_AUTHENTICATION_USE_CASE.PAYLOAD_NOT_MEET_DATA_TYPE_SPECIFICATION"
+      );
     }
   }
 }
